Fix up all the newly imported Teensy code

+++ motors/core/nonstd.c
@@ -28,7 +28,7 @@
  * SOFTWARE.
  */
 
-#include "avr_functions.h"
+#include "motors/core/nonstd.h"
 #include <string.h>
 #include <stdlib.h>
 #include <math.h>
@@ -171,7 +171,7 @@ char * dtostrf(float val, int width, unsigned int precision, char *buf)
 			s = fcvtf(val, newPrecision, &newDecimalPoint, &sign);
 
 			// if rounded up to new digit (e.g. 0.09 to 0.1), move decimal point
-			if (newDecimalPoint - decpt == precision + 1) decpt++;
+			if (newDecimalPoint - decpt == (int)precision + 1) decpt++;
 		}
 	}
